params,core/types: remove duplicate kzg version param

+++ params/protocol_params.go
@@ -166,7 +166,7 @@ const (
 	FieldElementsPerBlob = 4096 // each field element is 32 bytes
 
 	BlobVerificationGas      uint64 = 1800000
-	BlobCommitmentVersionKZG uint64 = 0x01
+	BlobCommitmentVersionKZG uint8  = 0x01
 	PointEvaluationGas       uint64 = 50000
 )
